Remove isodem feature
It removes the matplotlib dependency for `pyposeidon-base`. Keep it in utils for future consideration.
